% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{migparl_download_corpus}
\alias{migparl_download_corpus}
\title{Download corpora from Web Directory.}
\usage{
migparl_download_corpus(corpus = "MigParl",
  webdir = "https://polmine.sowi.uni-due.de/corpora/cwb/migparl",
  user = NULL, password = NULL, archive = FALSE, version = NULL)
}
\arguments{
\item{corpus}{tarball Name of the corpus to install.}

\item{webdir}{(web) directory where the tarball resides}

\item{user}{user info in case of password protected corpus}

\item{password}{password info in case of password protected corpus}

\item{archive}{logical; whether an older version of the corpus from the archive should be installed instead.}

\item{version}{if archive == TRUE, a corpus version must be specified.}
}
\description{
\code{migparl_download_corpus} will get a tarball with the indexed corpus
from a directory (web dir) and install the corpus into the MigParl package.
}
\examples{
\dontrun{
migparl_download_corpus(corpus = "MIGPARL", archive = TRUE, version = "2018-11-27")
}
}
